I added that benchmark file: bench/runbench.jl

There are major speedups with this little changes

Current master:

```
 ──────────────────────────────────────────────────────────────────
                           Time                   Allocations
                   ──────────────────────   ───────────────────────
 Tot / % measured:      85.9s / 100%            25.6GiB / 100%

 Section   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────
 c + s         10    17.3s  20.2%   1.73s   4.85GiB  19.0%   497MiB
   opt         10    9.82s  11.4%   982ms     0.00B  0.00%    0.00B
   copy        10    7.14s  8.32%   714ms   4.19GiB  16.4%   429MiB
   build       10    370ms  0.43%  37.0ms    676MiB  2.58%  67.6MiB
 bcs           10    17.2s  20.0%   1.72s   4.92GiB  19.2%   504MiB
   opt         10    16.8s  19.6%   1.68s   4.26GiB  16.6%   436MiB
   build       10    352ms  0.41%  35.2ms    677MiB  2.58%  67.7MiB
 cs            10    17.2s  20.0%   1.72s   4.92GiB  19.2%   504MiB
   opt         10    16.8s  19.6%   1.68s   4.26GiB  16.6%   436MiB
   build       10    370ms  0.43%  37.0ms    677MiB  2.58%  67.7MiB
 bcs + v       10    17.1s  19.9%   1.71s   5.45GiB  21.3%   558MiB
   opt         10    16.3s  19.0%   1.63s   4.26GiB  16.6%   436MiB
   build       10    719ms  0.84%  71.9ms   1.19GiB  4.66%   122MiB
 bc + s        10    16.4s  19.1%   1.64s   4.86GiB  19.0%   497MiB
   opt         10    8.91s  10.4%   891ms     0.00B  0.00%    0.00B
   copy        10    7.07s  8.24%   707ms   4.20GiB  16.4%   430MiB
   build       10    413ms  0.48%  41.3ms    676MiB  2.58%  67.6MiB
 data          10    677ms  0.79%  67.7ms    617MiB  2.35%  61.7MiB
 ──────────────────────────────────────────────────────────────────
```


After this PR:

```
 ──────────────────────────────────────────────────────────────────
                           Time                   Allocations
                   ──────────────────────   ───────────────────────
 Tot / % measured:      52.8s / 100%            8.91GiB / 100%

 Section   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────
 bcs + v       10    10.9s  20.7%   1.09s   2.11GiB  23.7%   216MiB
   opt         10    10.2s  19.3%   1.02s    938MiB  10.3%  93.8MiB
   build       10    744ms  1.41%  74.4ms   1.19GiB  13.4%   122MiB
 bcs           10    10.6s  20.0%   1.06s   1.58GiB  17.7%   161MiB
   opt         10    10.1s  19.1%   1.01s    938MiB  10.3%  93.8MiB
   build       10    454ms  0.86%  45.4ms    677MiB  7.43%  67.7MiB
 bc + s        10    10.3s  19.4%   1.03s   1.53GiB  17.1%   156MiB
   opt         10    8.31s  15.7%   831ms     0.00B  0.00%    0.00B
   copy        10    1.72s  3.25%   172ms    887MiB  9.73%  88.7MiB
   build       10    212ms  0.40%  21.2ms    676MiB  7.41%  67.6MiB
 cs            10    10.1s  19.2%   1.01s   1.58GiB  17.7%   161MiB
   opt         10    9.80s  18.6%   980ms    938MiB  10.3%  93.8MiB
   build       10    332ms  0.63%  33.2ms    677MiB  7.43%  67.7MiB
 c + s         10    10.1s  19.1%   1.01s   1.51GiB  17.0%   155MiB
   opt         10    8.15s  15.4%   815ms     0.00B  0.00%    0.00B
   copy        10    1.62s  3.07%   162ms    874MiB  9.58%  87.4MiB
   build       10    325ms  0.62%  32.5ms    676MiB  7.41%  67.6MiB
 data          10    789ms  1.49%  78.9ms    617MiB  6.76%  61.7MiB
 ──────────────────────────────────────────────────────────────────
```

cc @odow @mlubin @blegat