@joaquimg 2020-07-04
with the latest commit the bridging time is basically zero always:

```
 ──────────────────────────────────────────────────────────────────
                           Time                   Allocations
                   ──────────────────────   ───────────────────────
 Tot / % measured:      52.8s / 100%            8.91GiB / 100%

 Section   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────
 bcs + v       10    10.9s  20.7%   1.09s   2.11GiB  23.7%   216MiB
   opt         10    10.2s  19.3%   1.02s    938MiB  10.3%  93.8MiB
   build       10    744ms  1.41%  74.4ms   1.19GiB  13.4%   122MiB
 bcs           10    10.6s  20.0%   1.06s   1.58GiB  17.7%   161MiB
   opt         10    10.1s  19.1%   1.01s    938MiB  10.3%  93.8MiB
   build       10    454ms  0.86%  45.4ms    677MiB  7.43%  67.7MiB
 bc + s        10    10.3s  19.4%   1.03s   1.53GiB  17.1%   156MiB
   opt         10    8.31s  15.7%   831ms     0.00B  0.00%    0.00B
   copy        10    1.72s  3.25%   172ms    887MiB  9.73%  88.7MiB
   build       10    212ms  0.40%  21.2ms    676MiB  7.41%  67.6MiB
 cs            10    10.1s  19.2%   1.01s   1.58GiB  17.7%   161MiB
   opt         10    9.80s  18.6%   980ms    938MiB  10.3%  93.8MiB
   build       10    332ms  0.63%  33.2ms    677MiB  7.43%  67.7MiB
 c + s         10    10.1s  19.1%   1.01s   1.51GiB  17.0%   155MiB
   opt         10    8.15s  15.4%   815ms     0.00B  0.00%    0.00B
   copy        10    1.62s  3.07%   162ms    874MiB  9.58%  87.4MiB
   build       10    325ms  0.62%  32.5ms    676MiB  7.41%  67.6MiB
 data          10    789ms  1.49%  78.9ms    617MiB  6.76%  61.7MiB
 ──────────────────────────────────────────────────────────────────
```