@blegat 2020-07-04
Is `sparse` taking a lot of time ?
It seems we could create `A'` instead of `I, J, K`. For this, we need to call `MOIU.canonical(f)` as `SparseMatrixCSC` does not allow any duplicate and then add it as a new row of `A'` by adding an element to `colptr` and appending the new entries to `rowval` and `ǹzval`. Then we can do `A = copy(transpose(A'))` instead of `A = sparse(I, J, K, ...)`. One small detail, `SparseMatrixCSC` is immutable so we cannot modify `A'.n` and `A'.m`, so we probably need to just store `colptr`, `rowval` and `nzval` outside of the `SparseMatrixCSC` structure but the reasoning is the same.
Looking at the code of `SparseArrays.halfperm!`, it seems more efficient than `sparse`, this is due to the fact that `sparse` need to remove duplicates and sort while in this new approach this is already done in `MOIU.canonical`.
I was wondering about this thinking about the implementation of writing for MatrixOptInterface and improving the performance of SCS and the like (probably by using MatrixOptInterface) but I haven't tried it out yet.