@joaquimg 2020-07-04
> Is `sparse` taking a lot of time?

It does take some reasonable time.
Between 40 to 100ms, my guess is between 25 to 40% of copy time.

So yes, I think there is room for improvement.

Building colptr, rowval and nzval directly might be good.

I vote to merge this PR and experiment canonicalize +colptr, rowval and nzval of At is a new PR. Although, it seems that Clp would still require transposing as an extra step. On the other hand, other solvers take At instead of A, in which case it can be much better.

For Xpress, CPLEX and Gurobi, we could start with this copy_to function for the linear function and loop through the other constrains as today. Most of the time that we need high-performance loading is LP anyway. We can improve the other constraints loading as needed.