@joaquimg 2020-07-04
Last thing, YES, I like the idea of doing that in MatrixOptInterface.
Besides sharing code with Clp, Xpress, Bruobi, CPLEX, SCS etc.
MatrixOI can also be a key component for a good implementation of the differentiation GSOC.