@mtanneau 2020-07-04
In Tulip, I keep both A and A’ in memory, see
https://github.com/ds4dm/Tulip.jl/blob/master/src/Core/problemData.jl

A list of rows and columns of A is kept up-to-date. It allows for fast
column- and row-based operations, especially if the input is already sorted.